Feat: ProfileInfo 및 ProfilePosts 컴포넌트 분리
